+++ Engine/Container/ArrayPtr.h
@@ -152,7 +152,7 @@ public:
     /// Return pointer to the RefCount structure.
     RefCount* RefCountPtr() const { return refCount_; }
     /// Return hash value for HashSet & HashMap.
-    unsigned ToHash() const { return ((unsigned)ptr_) / sizeof(T); }
+    unsigned ToHash() const { return ((unsigned)(size_t)ptr_) / sizeof(T); }
     
 private:
     /// Prevent direct assignment from a shared array pointer of different type.
@@ -365,7 +365,7 @@ public:
     /// Return pointer to RefCount structure.
     RefCount* RefCountPtr() const { return refCount_; }
     /// Return hash value for HashSet & HashMap.
-    unsigned ToHash() const { return ((unsigned)ptr_) / sizeof(T); }
+    unsigned ToHash() const { return ((unsigned)(size_t)ptr_) / sizeof(T); }
     
 private:
     /// Prevent direct assignment from a weak array pointer of different type.
