+++ Engine/Container/Ptr.h
@@ -134,7 +134,7 @@ public:
     /// Return pointer to the RefCount structure.
     RefCount* RefCountPtr() const { return ptr_ ? ptr_->RefCountPtr() : 0; }
     /// Return hash value for HashSet & HashMap.
-    unsigned ToHash() const { return ((unsigned)ptr_) / sizeof(T); }
+    unsigned ToHash() const { return ((unsigned)(size_t)ptr_) / sizeof(T); }
     
 private:
     /// Prevent direct assignment from a shared pointer of another type.
@@ -363,7 +363,7 @@ public:
     /// Return pointer to the RefCount structure.
     RefCount* RefCountPtr() const { return refCount_; }
     /// Return hash value for HashSet & HashMap.
-    unsigned ToHash() const { return ((unsigned)ptr_) / sizeof(T); }
+    unsigned ToHash() const { return ((unsigned)(size_t)ptr_) / sizeof(T); }
     
 private:
     /// Prevent direct assignment from a weak pointer of different type.
